pyproject.toml
[project]
name = "python-ai-demo"
version = "0.1.0"
description = "AI Development Learning Roadmap - LLM Applications & AI Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Data & ML Fundamentals
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    # LLM APIs
    "openai>=1.58.1",
    "anthropic>=0.42.0",
    # LLM Frameworks
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "llama-index>=0.12.0",
    # Vector Databases & Embeddings
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    # Web & API
    "fastapi>=0.118.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.7.0", # Beautiful terminal output
    "tqdm>=4.66.0", # Progress bars
    "tiktoken>=0.5.0", # Token counting
    "llmlingua>=0.2.2",
    # Model Training (Phase 6)
    "torchvision>=0.19.0", # MNIST and image datasets
    "langchain-chroma>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "common",
    "phase1_foundations",
    "phase2_llm_fundamentals",
    "phase3_llm_applications",
    "phase4_ai_agents",
    "phase5_production",
    "phase6_model_customization"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # line length (handled by formatter)

[tool.mypy]
python_version = "3.11"
strict = false  # Start with relaxed, increase as you learn
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing"
pythonpath = ["."]
